refactor(cli): split main.go into separate command files
Reorganize CLI code for improved readability and maintainability:

- Create cmd/oastools/commands/ package with individual files per command
- common.go: Shared utilities (format validation, output helpers)
- parse.go, validate.go, join.go, convert.go, diff.go, generate.go, fix.go
- Simplify main.go to ~100 lines (was ~1400) with just command dispatch
- Add per-file unit tests following Go conventions
