+++ cmd/oastools/main.go
@@ -1,36 +1,12 @@
 package main
 
 import (
-	"encoding/json"
-	"errors"
-	"flag"
 	"fmt"
 	"os"
-	"path/filepath"
-	"time"
 
 	"github.com/erraggy/oastools"
-	"github.com/erraggy/oastools/converter"
-	"github.com/erraggy/oastools/differ"
-	"github.com/erraggy/oastools/fixer"
-	"github.com/erraggy/oastools/generator"
+	"github.com/erraggy/oastools/cmd/oastools/commands"
 	"github.com/erraggy/oastools/internal/cliutil"
-	"github.com/erraggy/oastools/joiner"
-	"github.com/erraggy/oastools/parser"
-	"github.com/erraggy/oastools/validator"
-	"go.yaml.in/yaml/v4"
-)
-
-// Output format constants
-const (
-	FormatText = "text"
-	FormatJSON = "json"
-	FormatYAML = "yaml"
-)
-
-// Special file path constant
-const (
-	StdinFilePath = "-"
 )
 
 func main() {
@@ -50,37 +26,37 @@ func main() {
 	case "help", "-h", "--help":
 		printUsage()
 	case "validate":
-		if err := handleValidate(os.Args[2:]); err != nil {
+		if err := commands.HandleValidate(os.Args[2:]); err != nil {
 			cliutil.Writef(os.Stderr, "Error: %v\n", err)
 			os.Exit(1)
 		}
 	case "parse":
-		if err := handleParse(os.Args[2:]); err != nil {
+		if err := commands.HandleParse(os.Args[2:]); err != nil {
 			cliutil.Writef(os.Stderr, "Error: %v\n", err)
 			os.Exit(1)
 		}
 	case "join":
-		if err := handleJoin(os.Args[2:]); err != nil {
+		if err := commands.HandleJoin(os.Args[2:]); err != nil {
 			cliutil.Writef(os.Stderr, "Error: %v\n", err)
 			os.Exit(1)
 		}
 	case "convert":
-		if err := handleConvert(os.Args[2:]); err != nil {
+		if err := commands.HandleConvert(os.Args[2:]); err != nil {
 			cliutil.Writef(os.Stderr, "Error: %v\n", err)
 			os.Exit(1)
 		}
 	case "diff":
-		if err := handleDiff(os.Args[2:]); err != nil {
+		if err := commands.HandleDiff(os.Args[2:]); err != nil {
 			cliutil.Writef(os.Stderr, "Error: %v\n", err)
 			os.Exit(1)
 		}
 	case "generate":
-		if err := handleGenerate(os.Args[2:]); err != nil {
+		if err := commands.HandleGenerate(os.Args[2:]); err != nil {
 			cliutil.Writef(os.Stderr, "Error: %v\n", err)
 			os.Exit(1)
 		}
 	case "fix":
-		if err := handleFix(os.Args[2:]); err != nil {
+		if err := commands.HandleFix(os.Args[2:]); err != nil {
 			cliutil.Writef(os.Stderr, "Error: %v\n", err)
 			os.Exit(1)
 		}
@@ -89,1264 +65,6 @@ func main() {
 		printUsage()
 		os.Exit(1)
 	}
-}
-
-// validateOutputFormat validates an output format and returns an error if invalid.
-func validateOutputFormat(format string) error {
-	if format != FormatText && format != FormatJSON && format != FormatYAML {
-		return fmt.Errorf("invalid format '%s'. Valid formats: %s, %s, %s", format, FormatText, FormatJSON, FormatYAML)
-	}
-	return nil
-}
-
-// outputStructured outputs data in the specified format (json or yaml) to stdout.
-// Returns an error if marshaling fails.
-func outputStructured(data any, format string) error {
-	var bytes []byte
-	var err error
-
-	switch format {
-	case FormatJSON:
-		bytes, err = json.MarshalIndent(data, "", "  ")
-	case FormatYAML:
-		bytes, err = yaml.Marshal(data)
-	default:
-		return fmt.Errorf("invalid format for structured output: %s", format)
-	}
-
-	if err != nil {
-		return fmt.Errorf("marshaling to %s: %w", format, err)
-	}
-
-	fmt.Println(string(bytes))
-	return nil
-}
-
-// validateCollisionStrategy validates a collision strategy name and returns an error if invalid.
-// The strategyName parameter is used in the error message (e.g., "path-strategy").
-func validateCollisionStrategy(strategyName, value string) error {
-	if value != "" && !joiner.IsValidStrategy(value) {
-		return fmt.Errorf("invalid %s '%s'. Valid strategies: %v", strategyName, value, joiner.ValidStrategies())
-	}
-	return nil
-}
-
-// parseFlags contains flags for the parse command
-type parseFlags struct {
-	resolveRefs       bool
-	resolveHTTPRefs   bool
-	insecure          bool
-	validateStructure bool
-	quiet             bool
-}
-
-// setupParseFlags creates and configures a FlagSet for the parse command.
-// Returns the FlagSet and a parseFlags struct with bound flag variables.
-func setupParseFlags() (*flag.FlagSet, *parseFlags) {
-	fs := flag.NewFlagSet("parse", flag.ContinueOnError)
-	flags := &parseFlags{}
-
-	fs.BoolVar(&flags.resolveRefs, "resolve-refs", false, "resolve external $ref references")
-	fs.BoolVar(&flags.resolveHTTPRefs, "resolve-http-refs", false, "resolve HTTP/HTTPS $ref URLs (requires --resolve-refs)")
-	fs.BoolVar(&flags.insecure, "insecure", false, "disable TLS certificate verification for HTTPS refs")
-	fs.BoolVar(&flags.validateStructure, "validate-structure", false, "validate document structure during parsing")
-	fs.BoolVar(&flags.quiet, "q", false, "quiet mode: only output the document, no diagnostic messages")
-	fs.BoolVar(&flags.quiet, "quiet", false, "quiet mode: only output the document, no diagnostic messages")
-
-	fs.Usage = func() {
-		output := fs.Output()
-		cliutil.Writef(output, "Usage: oastools parse [flags] <file|url|->\n\n")
-		cliutil.Writef(output, "Parse and output OpenAPI document structure.\n\n")
-		cliutil.Writef(output, "Flags:\n")
-		fs.PrintDefaults()
-		cliutil.Writef(output, "\nExamples:\n")
-		cliutil.Writef(output, "  oastools parse openapi.yaml\n")
-		cliutil.Writef(output, "  oastools parse --resolve-refs openapi.yaml\n")
-		cliutil.Writef(output, "  oastools parse --resolve-refs --resolve-http-refs https://example.com/api/openapi.yaml\n")
-		cliutil.Writef(output, "  oastools parse --resolve-refs --resolve-http-refs --insecure https://internal-server/api.yaml\n")
-		cliutil.Writef(output, "  oastools parse --validate-structure https://example.com/api/openapi.yaml\n")
-		cliutil.Writef(output, "  cat openapi.yaml | oastools parse -q -\n")
-		cliutil.Writef(output, "\nHTTP Reference Resolution:\n")
-		cliutil.Writef(output, "  --resolve-http-refs enables fetching and resolving $refs that point to HTTP/HTTPS URLs.\n")
-		cliutil.Writef(output, "  This is disabled by default for security (SSRF protection).\n")
-		cliutil.Writef(output, "  Use --insecure to skip TLS certificate verification for self-signed certs.\n")
-		cliutil.Writef(output, "\nPipelining:\n")
-		cliutil.Writef(output, "  - Use '-' as the file path to read from stdin\n")
-		cliutil.Writef(output, "  - Use --quiet/-q to suppress diagnostic output for pipelining\n")
-		cliutil.Writef(output, "\nExit Codes:\n")
-		cliutil.Writef(output, "  0    Parsing successful\n")
-		cliutil.Writef(output, "  1    Parsing failed or validation errors found (with --validate-structure)\n")
-	}
-
-	return fs, flags
-}
-
-func handleParse(args []string) error {
-	fs, flags := setupParseFlags()
-
-	if err := fs.Parse(args); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return nil
-		}
-		return err
-	}
-
-	if fs.NArg() != 1 {
-		fs.Usage()
-		return fmt.Errorf("parse command requires exactly one file path, URL, or '-' for stdin")
-	}
-
-	specPath := fs.Arg(0)
-
-	// Create parser with options
-	p := parser.New()
-	p.ResolveRefs = flags.resolveRefs
-	p.ResolveHTTPRefs = flags.resolveHTTPRefs
-	p.InsecureSkipVerify = flags.insecure
-	p.ValidateStructure = flags.validateStructure
-
-	// Parse the file, URL, or stdin
-	var result *parser.ParseResult
-	var err error
-
-	if specPath == StdinFilePath {
-		result, err = p.ParseReader(os.Stdin)
-		if err != nil {
-			return fmt.Errorf("parsing stdin: %w", err)
-		}
-	} else {
-		result, err = p.Parse(specPath)
-		if err != nil {
-			return fmt.Errorf("parsing file: %w", err)
-		}
-	}
-
-	// Always print errors to stderr, even in quiet mode (critical for debugging)
-	if len(result.Errors) > 0 {
-		cliutil.Writef(os.Stderr, "Validation Errors:\n")
-		for _, err := range result.Errors {
-			cliutil.Writef(os.Stderr, "  - %s\n", err)
-		}
-		cliutil.Writef(os.Stderr, "\n")
-		os.Exit(1)
-	}
-
-	// Print results (always to stderr to keep stdout clean for JSON output)
-	if !flags.quiet {
-		cliutil.Writef(os.Stderr, "OpenAPI Specification Parser\n")
-		cliutil.Writef(os.Stderr, "============================\n\n")
-		cliutil.Writef(os.Stderr, "oastools version: %s\n", oastools.Version())
-		if specPath == StdinFilePath {
-			cliutil.Writef(os.Stderr, "Specification: <stdin>\n")
-		} else {
-			cliutil.Writef(os.Stderr, "Specification: %s\n", specPath)
-		}
-		cliutil.Writef(os.Stderr, "OAS Version: %s\n", result.Version)
-		cliutil.Writef(os.Stderr, "Source Size: %s\n", parser.FormatBytes(result.SourceSize))
-		cliutil.Writef(os.Stderr, "Paths: %d\n", result.Stats.PathCount)
-		cliutil.Writef(os.Stderr, "Operations: %d\n", result.Stats.OperationCount)
-		cliutil.Writef(os.Stderr, "Schemas: %d\n", result.Stats.SchemaCount)
-		cliutil.Writef(os.Stderr, "Load Time: %v\n\n", result.LoadTime)
-
-		// Print warnings
-		if len(result.Warnings) > 0 {
-			cliutil.Writef(os.Stderr, "Warnings:\n")
-			for _, warning := range result.Warnings {
-				cliutil.Writef(os.Stderr, "  - %s\n", warning)
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-
-		// Print document info
-		if result.Document != nil {
-			switch doc := result.Document.(type) {
-			case *parser.OAS2Document:
-				cliutil.Writef(os.Stderr, "Document Type: OpenAPI 2.0 (Swagger)\n")
-				if doc.Info != nil {
-					cliutil.Writef(os.Stderr, "Title: %s\n", doc.Info.Title)
-					cliutil.Writef(os.Stderr, "Description: %s\n", doc.Info.Description)
-					cliutil.Writef(os.Stderr, "Version: %s\n", doc.Info.Version)
-				}
-				cliutil.Writef(os.Stderr, "Paths: %d\n", len(doc.Paths))
-
-			case *parser.OAS3Document:
-				cliutil.Writef(os.Stderr, "Document Type: OpenAPI 3.x\n")
-				if doc.Info != nil {
-					cliutil.Writef(os.Stderr, "Title: %s\n", doc.Info.Title)
-					if doc.Info.Summary != "" {
-						cliutil.Writef(os.Stderr, "Summary: %s\n", doc.Info.Summary)
-					}
-					cliutil.Writef(os.Stderr, "Description: %s\n", doc.Info.Description)
-					cliutil.Writef(os.Stderr, "Version: %s\n", doc.Info.Version)
-				}
-				cliutil.Writef(os.Stderr, "Servers: %d\n", len(doc.Servers))
-				cliutil.Writef(os.Stderr, "Paths: %d\n", len(doc.Paths))
-				if len(doc.Webhooks) > 0 {
-					cliutil.Writef(os.Stderr, "Webhooks: %d\n", len(doc.Webhooks))
-				}
-			}
-		}
-
-		cliutil.Writef(os.Stderr, "\n")
-		cliutil.Writef(os.Stderr, "Raw Data (JSON):\n")
-	}
-	jsonData, err := json.MarshalIndent(result.Data, "", "  ")
-	if err != nil {
-		return fmt.Errorf("marshaling to JSON: %w", err)
-	}
-	fmt.Println(string(jsonData))
-
-	if !flags.quiet {
-		cliutil.Writef(os.Stderr, "\nParsing completed successfully!\n")
-	}
-	return nil
-}
-
-// validateFlags contains flags for the validate command
-type validateFlags struct {
-	strict     bool
-	noWarnings bool
-	quiet      bool
-	format     string
-}
-
-// setupValidateFlags creates and configures a FlagSet for the validate command.
-// Returns the FlagSet and a validateFlags struct with bound flag variables.
-func setupValidateFlags() (*flag.FlagSet, *validateFlags) {
-	fs := flag.NewFlagSet("validate", flag.ContinueOnError)
-	flags := &validateFlags{}
-
-	fs.BoolVar(&flags.strict, "strict", false, "enable stricter validation beyond spec requirements")
-	fs.BoolVar(&flags.noWarnings, "no-warnings", false, "suppress warning messages (only show errors)")
-	fs.BoolVar(&flags.quiet, "q", false, "quiet mode: only output validation result, no diagnostic messages")
-	fs.BoolVar(&flags.quiet, "quiet", false, "quiet mode: only output validation result, no diagnostic messages")
-	fs.StringVar(&flags.format, "format", FormatText, "output format: text, json, or yaml")
-
-	fs.Usage = func() {
-		cliutil.Writef(fs.Output(), "Usage: oastools validate [flags] <file|url|->\n\n")
-		cliutil.Writef(fs.Output(), "Validate an OpenAPI specification file, URL, or stdin against the specification version it declares.\n\n")
-		cliutil.Writef(fs.Output(), "Flags:\n")
-		fs.PrintDefaults()
-		cliutil.Writef(fs.Output(), "\nOutput Formats:\n")
-		cliutil.Writef(fs.Output(), "  text (default)  Human-readable text output\n")
-		cliutil.Writef(fs.Output(), "  json            JSON format for programmatic processing\n")
-		cliutil.Writef(fs.Output(), "  yaml            YAML format for programmatic processing\n")
-		cliutil.Writef(fs.Output(), "\nExamples:\n")
-		cliutil.Writef(fs.Output(), "  oastools validate openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools validate https://example.com/api/openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools validate --strict api-spec.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools validate --no-warnings openapi.json\n")
-		cliutil.Writef(fs.Output(), "  cat openapi.yaml | oastools validate -q -\n")
-		cliutil.Writef(fs.Output(), "  oastools validate --format json openapi.yaml | jq '.valid'\n")
-		cliutil.Writef(fs.Output(), "\nPipelining:\n")
-		cliutil.Writef(fs.Output(), "  - Use '-' as the file path to read from stdin\n")
-		cliutil.Writef(fs.Output(), "  - Use --quiet/-q to suppress diagnostic output for pipelining\n")
-		cliutil.Writef(fs.Output(), "  - Use --format json/yaml for structured output that can be parsed\n")
-		cliutil.Writef(fs.Output(), "\nExit Codes:\n")
-		cliutil.Writef(fs.Output(), "  0    Validation successful\n")
-		cliutil.Writef(fs.Output(), "  1    Validation failed with errors\n")
-	}
-
-	return fs, flags
-}
-
-func handleValidate(args []string) error {
-	fs, flags := setupValidateFlags()
-
-	if err := fs.Parse(args); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return nil
-		}
-		return err
-	}
-
-	if fs.NArg() != 1 {
-		fs.Usage()
-		return fmt.Errorf("validate command requires exactly one file path, URL, or '-' for stdin")
-	}
-
-	specPath := fs.Arg(0)
-
-	// Validate format flag early to fail fast before expensive operations
-	if err := validateOutputFormat(flags.format); err != nil {
-		return err
-	}
-
-	// Create validator with options
-	v := validator.New()
-	v.StrictMode = flags.strict
-	v.IncludeWarnings = !flags.noWarnings
-
-	// Validate the file, URL, or stdin with timing
-	startTime := time.Now()
-	var result *validator.ValidationResult
-	var err error
-
-	if specPath == StdinFilePath {
-		// Read from stdin
-		p := parser.New()
-		parseResult, err := p.ParseReader(os.Stdin)
-		if err != nil {
-			return fmt.Errorf("parsing stdin: %w", err)
-		}
-		result, err = v.ValidateParsed(*parseResult)
-		if err != nil {
-			return fmt.Errorf("validating from stdin: %w", err)
-		}
-	} else {
-		result, err = v.Validate(specPath)
-		if err != nil {
-			return fmt.Errorf("validating file: %w", err)
-		}
-	}
-	totalTime := time.Since(startTime)
-
-	// Handle structured output formats
-	if flags.format == FormatJSON || flags.format == FormatYAML {
-		if err := outputStructured(result, flags.format); err != nil {
-			return err
-		}
-
-		// Exit with error if validation failed
-		if !result.Valid {
-			os.Exit(1)
-		}
-
-		return nil
-	}
-
-	// Text format output (original behavior)
-	// Print results (always to stderr to be consistent with parse and convert)
-	if !flags.quiet {
-		cliutil.Writef(os.Stderr, "OpenAPI Specification Validator\n")
-		cliutil.Writef(os.Stderr, "================================\n\n")
-		cliutil.Writef(os.Stderr, "oastools version: %s\n", oastools.Version())
-		if specPath == StdinFilePath {
-			cliutil.Writef(os.Stderr, "Specification: <stdin>\n")
-		} else {
-			cliutil.Writef(os.Stderr, "Specification: %s\n", specPath)
-		}
-		cliutil.Writef(os.Stderr, "OAS Version: %s\n", result.Version)
-		cliutil.Writef(os.Stderr, "Source Size: %s\n", parser.FormatBytes(result.SourceSize))
-		cliutil.Writef(os.Stderr, "Paths: %d\n", result.Stats.PathCount)
-		cliutil.Writef(os.Stderr, "Operations: %d\n", result.Stats.OperationCount)
-		cliutil.Writef(os.Stderr, "Schemas: %d\n", result.Stats.SchemaCount)
-		cliutil.Writef(os.Stderr, "Load Time: %v\n", result.LoadTime)
-		cliutil.Writef(os.Stderr, "Total Time: %v\n\n", totalTime)
-
-		// Print errors
-		if len(result.Errors) > 0 {
-			cliutil.Writef(os.Stderr, "Errors (%d):\n", result.ErrorCount)
-			for _, err := range result.Errors {
-				cliutil.Writef(os.Stderr, "  %s\n", err.String())
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-
-		// Print warnings
-		if len(result.Warnings) > 0 {
-			cliutil.Writef(os.Stderr, "Warnings (%d):\n", result.WarningCount)
-			for _, warning := range result.Warnings {
-				cliutil.Writef(os.Stderr, "  %s\n", warning.String())
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-	}
-
-	// Print summary (only in non-quiet mode to respect --quiet flag)
-	if !flags.quiet {
-		if result.Valid {
-			cliutil.Writef(os.Stderr, "✓ Validation passed")
-			if result.WarningCount > 0 {
-				cliutil.Writef(os.Stderr, " with %d warning(s)", result.WarningCount)
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		} else {
-			cliutil.Writef(os.Stderr, "✗ Validation failed: %d error(s)", result.ErrorCount)
-			if result.WarningCount > 0 {
-				cliutil.Writef(os.Stderr, ", %d warning(s)", result.WarningCount)
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-	}
-
-	// Exit with error if validation failed
-	if !result.Valid {
-		os.Exit(1)
-	}
-
-	return nil
-}
-
-// joinFlags contains flags for the join command
-type joinFlags struct {
-	output            string
-	pathStrategy      string
-	schemaStrategy    string
-	componentStrategy string
-	noMergeArrays     bool
-	noDedupTags       bool
-	quiet             bool
-}
-
-// setupJoinFlags creates and configures a FlagSet for the join command.
-// Returns the FlagSet and a joinFlags struct with bound flag variables.
-func setupJoinFlags() (*flag.FlagSet, *joinFlags) {
-	fs := flag.NewFlagSet("join", flag.ContinueOnError)
-	flags := &joinFlags{}
-
-	fs.StringVar(&flags.output, "o", "", "output file path (default: stdout)")
-	fs.StringVar(&flags.output, "output", "", "output file path (default: stdout)")
-	fs.StringVar(&flags.pathStrategy, "path-strategy", "", "collision strategy for paths (accept-left, accept-right, fail, fail-on-paths)")
-	fs.StringVar(&flags.schemaStrategy, "schema-strategy", "", "collision strategy for schemas/definitions")
-	fs.StringVar(&flags.componentStrategy, "component-strategy", "", "collision strategy for other components")
-	fs.BoolVar(&flags.noMergeArrays, "no-merge-arrays", false, "don't merge arrays (servers, security, etc.)")
-	fs.BoolVar(&flags.noDedupTags, "no-dedup-tags", false, "don't deduplicate tags by name")
-	fs.BoolVar(&flags.quiet, "q", false, "quiet mode: suppress diagnostic messages (for pipelining)")
-	fs.BoolVar(&flags.quiet, "quiet", false, "quiet mode: suppress diagnostic messages (for pipelining)")
-
-	fs.Usage = func() {
-		cliutil.Writef(fs.Output(), "Usage: oastools join [flags] <file1> <file2> [file3...]\n\n")
-		cliutil.Writef(fs.Output(), "Join multiple OpenAPI specification files into a single document.\n\n")
-		cliutil.Writef(fs.Output(), "Flags:\n")
-		fs.PrintDefaults()
-		cliutil.Writef(fs.Output(), "\nCollision Strategies:\n")
-		cliutil.Writef(fs.Output(), "  accept-left      Keep the first value when collisions occur\n")
-		cliutil.Writef(fs.Output(), "  accept-right     Keep the last value when collisions occur (overwrite)\n")
-		cliutil.Writef(fs.Output(), "  fail             Fail with an error on any collision\n")
-		cliutil.Writef(fs.Output(), "  fail-on-paths    Fail only on path collisions, allow schema collisions\n")
-		cliutil.Writef(fs.Output(), "\nExamples:\n")
-		cliutil.Writef(fs.Output(), "  oastools join -o merged.yaml base.yaml extensions.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools join --path-strategy accept-left -o api.yaml spec1.yaml spec2.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools join --schema-strategy accept-right -o output.yaml api1.yaml api2.yaml api3.yaml\n")
-		cliutil.Writef(fs.Output(), "\nPipelining:\n")
-		cliutil.Writef(fs.Output(), "  oastools join -q base.yaml ext.yaml | oastools validate -q -\n")
-		cliutil.Writef(fs.Output(), "  oastools join -q spec1.yaml spec2.yaml | oastools convert -q -t 3.1.0 -\n")
-		cliutil.Writef(fs.Output(), "\nNotes:\n")
-		cliutil.Writef(fs.Output(), "  - All input files must be the same major OAS version (2.0 or 3.x)\n")
-		cliutil.Writef(fs.Output(), "  - The output will use the version of the first input file\n")
-		cliutil.Writef(fs.Output(), "  - Info section is taken from the first document by default\n")
-		cliutil.Writef(fs.Output(), "  - When -o is specified, file is written with restrictive permissions (0600)\n")
-	}
-
-	return fs, flags
-}
-
-func handleJoin(args []string) error {
-	fs, flags := setupJoinFlags()
-
-	if err := fs.Parse(args); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return nil
-		}
-		return err
-	}
-
-	if fs.NArg() < 2 {
-		fs.Usage()
-		return fmt.Errorf("join command requires at least 2 input files")
-	}
-
-	filePaths := fs.Args()
-
-	// Build configuration
-	config := joiner.DefaultConfig()
-	config.MergeArrays = !flags.noMergeArrays
-	config.DeduplicateTags = !flags.noDedupTags
-
-	// Validate and parse strategy flags
-	if err := validateCollisionStrategy("path-strategy", flags.pathStrategy); err != nil {
-		return err
-	}
-	if err := validateCollisionStrategy("schema-strategy", flags.schemaStrategy); err != nil {
-		return err
-	}
-	if err := validateCollisionStrategy("component-strategy", flags.componentStrategy); err != nil {
-		return err
-	}
-
-	// Apply validated strategies to config
-	if flags.pathStrategy != "" {
-		config.PathStrategy = joiner.CollisionStrategy(flags.pathStrategy)
-	}
-	if flags.schemaStrategy != "" {
-		config.SchemaStrategy = joiner.CollisionStrategy(flags.schemaStrategy)
-	}
-	if flags.componentStrategy != "" {
-		config.ComponentStrategy = joiner.CollisionStrategy(flags.componentStrategy)
-	}
-
-	// Validate output path before joining (only when writing to file)
-	if flags.output != "" {
-		if err := validateOutputPath(flags.output, filePaths); err != nil {
-			return err
-		}
-	}
-
-	// Create joiner and execute with timing
-	startTime := time.Now()
-	j := joiner.New(config)
-	result, err := j.Join(filePaths)
-	if err != nil {
-		return fmt.Errorf("joining specifications: %w", err)
-	}
-	totalTime := time.Since(startTime)
-
-	// Print diagnostic messages (to stderr to keep stdout clean for pipelining)
-	if !flags.quiet {
-		cliutil.Writef(os.Stderr, "OpenAPI Specification Joiner\n")
-		cliutil.Writef(os.Stderr, "============================\n\n")
-		cliutil.Writef(os.Stderr, "oastools version: %s\n", oastools.Version())
-		cliutil.Writef(os.Stderr, "Successfully joined %d specification files\n", len(filePaths))
-		if flags.output != "" {
-			cliutil.Writef(os.Stderr, "Output: %s\n", flags.output)
-		} else {
-			cliutil.Writef(os.Stderr, "Output: <stdout>\n")
-		}
-		cliutil.Writef(os.Stderr, "OAS Version: %s\n", result.Version)
-		cliutil.Writef(os.Stderr, "Paths: %d\n", result.Stats.PathCount)
-		cliutil.Writef(os.Stderr, "Operations: %d\n", result.Stats.OperationCount)
-		cliutil.Writef(os.Stderr, "Schemas: %d\n", result.Stats.SchemaCount)
-		cliutil.Writef(os.Stderr, "Total Time: %v\n\n", totalTime)
-
-		if result.CollisionCount > 0 {
-			cliutil.Writef(os.Stderr, "Collisions resolved: %d\n\n", result.CollisionCount)
-		}
-
-		if len(result.Warnings) > 0 {
-			cliutil.Writef(os.Stderr, "Warnings (%d):\n", len(result.Warnings))
-			for _, warning := range result.Warnings {
-				cliutil.Writef(os.Stderr, "  - %s\n", warning)
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-
-		cliutil.Writef(os.Stderr, "✓ Join completed successfully!\n")
-	}
-
-	// Write output
-	if flags.output != "" {
-		// Write to file
-		err = j.WriteResult(result, flags.output)
-		if err != nil {
-			return fmt.Errorf("writing output file: %w", err)
-		}
-		if !flags.quiet {
-			cliutil.Writef(os.Stderr, "\nOutput written to: %s\n", flags.output)
-		}
-	} else {
-		// Write to stdout
-		data, err := marshalDocument(result.Document, result.SourceFormat)
-		if err != nil {
-			return fmt.Errorf("marshaling joined document: %w", err)
-		}
-		if _, err = os.Stdout.Write(data); err != nil {
-			return fmt.Errorf("writing joined document to stdout: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// validateOutputPath checks if the output path is safe to write to
-func validateOutputPath(outputPath string, inputPaths []string) error {
-	// Get absolute path of output file
-	absOutputPath, err := filepath.Abs(outputPath)
-	if err != nil {
-		return fmt.Errorf("invalid output path: %w", err)
-	}
-
-	// Check if output file would overwrite any input files
-	for _, inputPath := range inputPaths {
-		absInputPath, err := filepath.Abs(inputPath)
-		if err != nil {
-			return fmt.Errorf("invalid input path %s: %w", inputPath, err)
-		}
-
-		if absOutputPath == absInputPath {
-			return fmt.Errorf("output file %s would overwrite input file %s", outputPath, inputPath)
-		}
-	}
-
-	// Check if output file already exists and warn (but don't error)
-	if _, err := os.Stat(outputPath); err == nil {
-		cliutil.Writef(os.Stderr, "Warning: output file %s already exists and will be overwritten\n", outputPath)
-	}
-
-	return nil
-}
-
-// convertFlags contains flags for the convert command
-type convertFlags struct {
-	target     string
-	output     string
-	strict     bool
-	noWarnings bool
-	quiet      bool
-}
-
-// setupConvertFlags creates and configures a FlagSet for the convert command.
-// Returns the FlagSet and a convertFlags struct with bound flag variables.
-func setupConvertFlags() (*flag.FlagSet, *convertFlags) {
-	fs := flag.NewFlagSet("convert", flag.ContinueOnError)
-	flags := &convertFlags{}
-
-	fs.StringVar(&flags.target, "t", "", "target OAS version (e.g., \"3.0.3\", \"2.0\", \"3.1.0\") (required)")
-	fs.StringVar(&flags.target, "target", "", "target OAS version (e.g., \"3.0.3\", \"2.0\", \"3.1.0\") (required)")
-	fs.StringVar(&flags.output, "o", "", "output file path (default: stdout)")
-	fs.StringVar(&flags.output, "output", "", "output file path (default: stdout)")
-	fs.BoolVar(&flags.strict, "strict", false, "fail on any conversion issues (even warnings)")
-	fs.BoolVar(&flags.noWarnings, "no-warnings", false, "suppress warning and info messages")
-	fs.BoolVar(&flags.quiet, "q", false, "quiet mode: only output the document, no diagnostic messages")
-	fs.BoolVar(&flags.quiet, "quiet", false, "quiet mode: only output the document, no diagnostic messages")
-
-	fs.Usage = func() {
-		cliutil.Writef(fs.Output(), "Usage: oastools convert [flags] <file|url|->\n\n")
-		cliutil.Writef(fs.Output(), "Convert an OpenAPI specification from one version to another.\n\n")
-		cliutil.Writef(fs.Output(), "Flags:\n")
-		fs.PrintDefaults()
-		cliutil.Writef(fs.Output(), "\nSupported Conversions:\n")
-		cliutil.Writef(fs.Output(), "  - OAS 2.0 → OAS 3.x (3.0.0 through 3.2.0)\n")
-		cliutil.Writef(fs.Output(), "  - OAS 3.x → OAS 2.0\n")
-		cliutil.Writef(fs.Output(), "  - OAS 3.x → OAS 3.y (version updates)\n")
-		cliutil.Writef(fs.Output(), "\nExamples:\n")
-		cliutil.Writef(fs.Output(), "  oastools convert -t 3.0.3 swagger.yaml -o openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools convert -t 3.0.3 https://example.com/swagger.yaml -o openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools convert -t 2.0 openapi-v3.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools convert --strict -t 3.1.0 swagger.yaml -o openapi-v3.yaml\n")
-		cliutil.Writef(fs.Output(), "  cat swagger.yaml | oastools convert -q -t 3.0.3 - > openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "\nPipelining:\n")
-		cliutil.Writef(fs.Output(), "  - Use '-' as the file path to read from stdin\n")
-		cliutil.Writef(fs.Output(), "  - Use --quiet/-q to suppress diagnostic output for pipelining\n")
-		cliutil.Writef(fs.Output(), "\nNotes:\n")
-		cliutil.Writef(fs.Output(), "  - Critical issues indicate features that cannot be converted (data loss)\n")
-		cliutil.Writef(fs.Output(), "  - Warnings indicate lossy conversions or best-effort transformations\n")
-		cliutil.Writef(fs.Output(), "  - Info messages provide context about conversion choices\n")
-		cliutil.Writef(fs.Output(), "  - Always validate converted documents before deployment\n")
-		cliutil.Writef(fs.Output(), "\nExit Codes:\n")
-		cliutil.Writef(fs.Output(), "  0    Conversion successful\n")
-		cliutil.Writef(fs.Output(), "  1    Conversion failed or critical issues found (in --strict mode)\n")
-	}
-
-	return fs, flags
-}
-
-func handleConvert(args []string) error {
-	fs, flags := setupConvertFlags()
-
-	if err := fs.Parse(args); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return nil
-		}
-		return err
-	}
-
-	if fs.NArg() != 1 {
-		fs.Usage()
-		return fmt.Errorf("convert command requires exactly one file path, URL, or '-' for stdin")
-	}
-
-	specPath := fs.Arg(0)
-
-	if flags.target == "" {
-		fs.Usage()
-		return fmt.Errorf("target version is required (use -t or --target)")
-	}
-
-	// Create converter with options
-	c := converter.New()
-	c.StrictMode = flags.strict
-	c.IncludeInfo = !flags.noWarnings
-
-	// Convert the file, URL, or stdin with timing
-	startTime := time.Now()
-	var result *converter.ConversionResult
-	var err error
-
-	if specPath == StdinFilePath {
-		// Read from stdin
-		p := parser.New()
-		parseResult, err := p.ParseReader(os.Stdin)
-		if err != nil {
-			return fmt.Errorf("parsing stdin: %w", err)
-		}
-		result, err = c.ConvertParsed(*parseResult, flags.target)
-		if err != nil {
-			return fmt.Errorf("converting from stdin: %w", err)
-		}
-	} else {
-		result, err = c.Convert(specPath, flags.target)
-		if err != nil {
-			return fmt.Errorf("converting file: %w", err)
-		}
-	}
-	totalTime := time.Since(startTime)
-
-	// Print results (to stderr in quiet mode)
-	if !flags.quiet {
-		cliutil.Writef(os.Stderr, "OpenAPI Specification Converter\n")
-		cliutil.Writef(os.Stderr, "===============================\n\n")
-		cliutil.Writef(os.Stderr, "oastools version: %s\n", oastools.Version())
-		if specPath == StdinFilePath {
-			cliutil.Writef(os.Stderr, "Specification: <stdin>\n")
-		} else {
-			cliutil.Writef(os.Stderr, "Specification: %s\n", specPath)
-		}
-		cliutil.Writef(os.Stderr, "Source Version: %s\n", result.SourceVersion)
-		cliutil.Writef(os.Stderr, "Target Version: %s\n", result.TargetVersion)
-		cliutil.Writef(os.Stderr, "Source Size: %s\n", parser.FormatBytes(result.SourceSize))
-		cliutil.Writef(os.Stderr, "Paths: %d\n", result.Stats.PathCount)
-		cliutil.Writef(os.Stderr, "Operations: %d\n", result.Stats.OperationCount)
-		cliutil.Writef(os.Stderr, "Schemas: %d\n", result.Stats.SchemaCount)
-		cliutil.Writef(os.Stderr, "Load Time: %v\n", result.LoadTime)
-		cliutil.Writef(os.Stderr, "Total Time: %v\n\n", totalTime)
-
-		// Print issues
-		if len(result.Issues) > 0 {
-			cliutil.Writef(os.Stderr, "Conversion Issues (%d):\n", len(result.Issues))
-			for _, issue := range result.Issues {
-				cliutil.Writef(os.Stderr, "  %s\n", issue.String())
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-
-		// Print summary
-		if result.Success {
-			cliutil.Writef(os.Stderr, "✓ Conversion successful")
-			if result.InfoCount > 0 || result.WarningCount > 0 {
-				cliutil.Writef(os.Stderr, " (%d info, %d warnings)", result.InfoCount, result.WarningCount)
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		} else {
-			cliutil.Writef(os.Stderr, "✗ Conversion completed with %d critical issue(s)", result.CriticalCount)
-			if result.WarningCount > 0 {
-				cliutil.Writef(os.Stderr, ", %d warning(s)", result.WarningCount)
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-	}
-
-	// Write output
-	data, err := marshalDocument(result.Document, result.SourceFormat)
-	if err != nil {
-		return fmt.Errorf("marshaling converted document: %w", err)
-	}
-
-	if flags.output != "" {
-		if err := os.WriteFile(flags.output, data, 0600); err != nil {
-			return fmt.Errorf("writing output file: %w", err)
-		}
-		if !flags.quiet {
-			cliutil.Writef(os.Stderr, "\nOutput written to: %s\n", flags.output)
-		}
-	} else {
-		// Write to stdout
-		if _, err = os.Stdout.Write(data); err != nil {
-			return fmt.Errorf("writing converted document to stdout: %w", err)
-		}
-	}
-
-	// Exit with error if conversion failed
-	if !result.Success {
-		os.Exit(1)
-	}
-
-	return nil
-}
-
-// marshalDocument marshals a document to bytes in the specified format
-func marshalDocument(doc any, format parser.SourceFormat) ([]byte, error) {
-	if format == parser.SourceFormatJSON {
-		return json.MarshalIndent(doc, "", "  ")
-	}
-	return yaml.Marshal(doc)
-}
-
-// diffFlags contains flags for the diff command
-type diffFlags struct {
-	breaking bool
-	noInfo   bool
-	format   string
-}
-
-// setupDiffFlags creates and configures a FlagSet for the diff command.
-// Returns the FlagSet and a diffFlags struct with bound flag variables.
-func setupDiffFlags() (*flag.FlagSet, *diffFlags) {
-	fs := flag.NewFlagSet("diff", flag.ContinueOnError)
-	flags := &diffFlags{}
-
-	fs.BoolVar(&flags.breaking, "breaking", false, "enable breaking change detection and categorization")
-	fs.BoolVar(&flags.noInfo, "no-info", false, "exclude informational changes from output")
-	fs.StringVar(&flags.format, "format", FormatText, "output format: text, json, or yaml")
-
-	fs.Usage = func() {
-		cliutil.Writef(fs.Output(), "Usage: oastools diff [flags] <source> <target>\n\n")
-		cliutil.Writef(fs.Output(), "Compare two OpenAPI specification files or URLs and report differences.\n\n")
-		cliutil.Writef(fs.Output(), "Flags:\n")
-		fs.PrintDefaults()
-		cliutil.Writef(fs.Output(), "\nOutput Formats:\n")
-		cliutil.Writef(fs.Output(), "  text (default)  Human-readable text output\n")
-		cliutil.Writef(fs.Output(), "  json            JSON format for programmatic processing\n")
-		cliutil.Writef(fs.Output(), "  yaml            YAML format for programmatic processing\n")
-		cliutil.Writef(fs.Output(), "\nModes:\n")
-		cliutil.Writef(fs.Output(), "  Default (Simple):\n")
-		cliutil.Writef(fs.Output(), "    Reports all semantic differences between specifications without\n")
-		cliutil.Writef(fs.Output(), "    categorizing them by severity or breaking change impact.\n\n")
-		cliutil.Writef(fs.Output(), "  --breaking (Breaking Change Detection):\n")
-		cliutil.Writef(fs.Output(), "    Categorizes changes by severity and identifies breaking API changes:\n")
-		cliutil.Writef(fs.Output(), "    - Critical: Removed endpoints or operations\n")
-		cliutil.Writef(fs.Output(), "    - Error:    Removed required parameters, incompatible type changes\n")
-		cliutil.Writef(fs.Output(), "    - Warning:  Deprecated operations, added required fields\n")
-		cliutil.Writef(fs.Output(), "    - Info:     Additions, relaxed constraints, documentation updates\n")
-		cliutil.Writef(fs.Output(), "\nExamples:\n")
-		cliutil.Writef(fs.Output(), "  oastools diff api-v1.yaml api-v2.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools diff --breaking api-v1.yaml api-v2.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools diff --breaking --no-info old.yaml new.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools diff --format json --breaking api-v1.yaml api-v2.yaml | jq '.HasBreakingChanges'\n")
-		cliutil.Writef(fs.Output(), "  oastools diff https://example.com/api/v1.yaml https://example.com/api/v2.yaml\n")
-		cliutil.Writef(fs.Output(), "\nExit Status:\n")
-		cliutil.Writef(fs.Output(), "  0    No differences found (or no breaking changes in --breaking mode)\n")
-		cliutil.Writef(fs.Output(), "  1    Differences found (or breaking changes found in --breaking mode)\n")
-		cliutil.Writef(fs.Output(), "\nNotes:\n")
-		cliutil.Writef(fs.Output(), "  - Both specifications must be valid OpenAPI documents\n")
-		cliutil.Writef(fs.Output(), "  - Cross-version comparison (2.0 vs 3.x) is supported with limitations\n")
-		cliutil.Writef(fs.Output(), "  - Breaking change detection helps identify backward compatibility issues\n")
-	}
-
-	return fs, flags
-}
-
-func handleDiff(args []string) error {
-	fs, flags := setupDiffFlags()
-
-	if err := fs.Parse(args); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return nil
-		}
-		return err
-	}
-
-	if fs.NArg() != 2 {
-		fs.Usage()
-		return fmt.Errorf("diff command requires exactly two file paths or URLs")
-	}
-
-	sourcePath := fs.Arg(0)
-	targetPath := fs.Arg(1)
-
-	// Validate format flag
-	if err := validateOutputFormat(flags.format); err != nil {
-		return err
-	}
-
-	// Create differ with options
-	d := differ.New()
-	if flags.breaking {
-		d.Mode = differ.ModeBreaking
-	} else {
-		d.Mode = differ.ModeSimple
-	}
-	d.IncludeInfo = !flags.noInfo
-
-	// Diff the files with timing
-	startTime := time.Now()
-	result, err := d.Diff(sourcePath, targetPath)
-	totalTime := time.Since(startTime)
-	if err != nil {
-		return fmt.Errorf("comparing specifications: %w", err)
-	}
-
-	// Handle structured output formats
-	if flags.format == FormatJSON || flags.format == FormatYAML {
-		if err := outputStructured(result, flags.format); err != nil {
-			return err
-		}
-
-		// Exit with error if breaking changes found (in breaking mode)
-		if flags.breaking && result.HasBreakingChanges {
-			os.Exit(1)
-		}
-
-		return nil
-	}
-
-	// Text format output (original behavior)
-	// Print results
-	fmt.Printf("OpenAPI Specification Diff\n")
-	fmt.Printf("==========================\n\n")
-	fmt.Printf("oastools version: %s\n", oastools.Version())
-	fmt.Printf("Source: %s (%s)\n", sourcePath, result.SourceVersion)
-	fmt.Printf("Target: %s (%s)\n", targetPath, result.TargetVersion)
-	fmt.Printf("Total Time: %v\n\n", totalTime)
-
-	if len(result.Changes) == 0 {
-		fmt.Println("✓ No differences found - specifications are identical")
-		return nil
-	}
-
-	// Print changes grouped by category if in breaking mode
-	if flags.breaking {
-		// Group changes by category
-		categories := make(map[differ.ChangeCategory][]differ.Change)
-		for _, change := range result.Changes {
-			categories[change.Category] = append(categories[change.Category], change)
-		}
-
-		// Print each category
-		categoryOrder := []differ.ChangeCategory{
-			differ.CategoryEndpoint,
-			differ.CategoryOperation,
-			differ.CategoryParameter,
-			differ.CategoryRequestBody,
-			differ.CategoryResponse,
-			differ.CategorySchema,
-			differ.CategorySecurity,
-			differ.CategoryServer,
-			differ.CategoryInfo,
-		}
-
-		for _, category := range categoryOrder {
-			changes := categories[category]
-			if len(changes) == 0 {
-				continue
-			}
-
-			fmt.Printf("%s Changes (%d):\n", category, len(changes))
-			for _, change := range changes {
-				fmt.Printf("  %s\n", change.String())
-			}
-			fmt.Println()
-		}
-
-		// Print summary
-		fmt.Printf("Summary:\n")
-		fmt.Printf("  Total changes: %d\n", len(result.Changes))
-		if result.HasBreakingChanges {
-			fmt.Printf("  ⚠️  Breaking changes: %d\n", result.BreakingCount)
-		} else {
-			fmt.Printf("  ✓ Breaking changes: 0\n")
-		}
-		fmt.Printf("  Warnings: %d\n", result.WarningCount)
-		if d.IncludeInfo {
-			fmt.Printf("  Info: %d\n", result.InfoCount)
-		}
-
-		// Exit with error if breaking changes found
-		if result.HasBreakingChanges {
-			os.Exit(1)
-		}
-	} else {
-		// Simple mode - just print all changes
-		fmt.Printf("Changes (%d):\n", len(result.Changes))
-		for _, change := range result.Changes {
-			fmt.Printf("  %s\n", change.String())
-		}
-	}
-
-	return nil
-}
-
-// generateFlags contains flags for the generate command
-type generateFlags struct {
-	output       string
-	packageName  string
-	client       bool
-	server       bool
-	types        bool
-	noPointers   bool
-	noValidation bool
-	strict       bool
-	noWarnings   bool
-}
-
-// setupGenerateFlags creates and configures a FlagSet for the generate command.
-// Returns the FlagSet and a generateFlags struct with bound flag variables.
-func setupGenerateFlags() (*flag.FlagSet, *generateFlags) {
-	fs := flag.NewFlagSet("generate", flag.ContinueOnError)
-	flags := &generateFlags{}
-
-	fs.StringVar(&flags.output, "o", "", "output directory for generated files (required)")
-	fs.StringVar(&flags.output, "output", "", "output directory for generated files (required)")
-	fs.StringVar(&flags.packageName, "p", "api", "Go package name for generated code")
-	fs.StringVar(&flags.packageName, "package", "api", "Go package name for generated code")
-	fs.BoolVar(&flags.client, "client", false, "generate HTTP client code")
-	fs.BoolVar(&flags.server, "server", false, "generate server interface code")
-	fs.BoolVar(&flags.types, "types", true, "generate type definitions from schemas")
-	fs.BoolVar(&flags.noPointers, "no-pointers", false, "don't use pointer types for optional fields")
-	fs.BoolVar(&flags.noValidation, "no-validation", false, "don't include validation tags")
-	fs.BoolVar(&flags.strict, "strict", false, "fail on any generation issues (even warnings)")
-	fs.BoolVar(&flags.noWarnings, "no-warnings", false, "suppress warning and info messages")
-
-	fs.Usage = func() {
-		cliutil.Writef(fs.Output(), "Usage: oastools generate [flags] <file|url>\n\n")
-		cliutil.Writef(fs.Output(), "Generate Go code from an OpenAPI specification.\n\n")
-		cliutil.Writef(fs.Output(), "Flags:\n")
-		fs.PrintDefaults()
-		cliutil.Writef(fs.Output(), "\nExamples:\n")
-		cliutil.Writef(fs.Output(), "  oastools generate --client -o ./client openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools generate --server -o ./server -p myapi openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools generate --client --server -o ./api petstore.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools generate --types -o ./models https://example.com/api/openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "\nNotes:\n")
-		cliutil.Writef(fs.Output(), "  - At least one of --client, --server, or --types must be enabled\n")
-		cliutil.Writef(fs.Output(), "  - Types are always generated when --client or --server is enabled\n")
-		cliutil.Writef(fs.Output(), "  - Generated code uses Go idioms and best practices\n")
-		cliutil.Writef(fs.Output(), "  - Server interface is framework-agnostic\n")
-	}
-
-	return fs, flags
-}
-
-func handleGenerate(args []string) error {
-	fs, flags := setupGenerateFlags()
-
-	if err := fs.Parse(args); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return nil
-		}
-		return err
-	}
-
-	if fs.NArg() != 1 {
-		fs.Usage()
-		return fmt.Errorf("generate command requires exactly one file path or URL")
-	}
-
-	specPath := fs.Arg(0)
-
-	if flags.output == "" {
-		fs.Usage()
-		return fmt.Errorf("output directory is required (use -o or --output)")
-	}
-
-	// Ensure at least one generation mode is enabled
-	if !flags.client && !flags.server && !flags.types {
-		fs.Usage()
-		return fmt.Errorf("at least one of --client, --server, or --types must be enabled")
-	}
-
-	// Create generator with options
-	g := generator.New()
-	g.PackageName = flags.packageName
-	g.GenerateClient = flags.client
-	g.GenerateServer = flags.server
-	g.GenerateTypes = flags.types || flags.client || flags.server
-	g.UsePointers = !flags.noPointers
-	g.IncludeValidation = !flags.noValidation
-	g.StrictMode = flags.strict
-	g.IncludeInfo = !flags.noWarnings
-
-	// Generate the code with timing
-	startTime := time.Now()
-	result, err := g.Generate(specPath)
-	totalTime := time.Since(startTime)
-	if err != nil {
-		return fmt.Errorf("generating code: %w", err)
-	}
-
-	// Print results
-	fmt.Printf("OpenAPI Code Generator\n")
-	fmt.Printf("=====================\n\n")
-	fmt.Printf("oastools version: %s\n", oastools.Version())
-	fmt.Printf("Specification: %s\n", specPath)
-	fmt.Printf("OAS Version: %s\n", result.SourceVersion)
-	fmt.Printf("Source Size: %s\n", parser.FormatBytes(result.SourceSize))
-	fmt.Printf("Package: %s\n", result.PackageName)
-	fmt.Printf("Types: %d\n", result.GeneratedTypes)
-	fmt.Printf("Operations: %d\n", result.GeneratedOperations)
-	fmt.Printf("Total Time: %v\n\n", totalTime)
-
-	// Print issues
-	if len(result.Issues) > 0 {
-		fmt.Printf("Generation Issues (%d):\n", len(result.Issues))
-		for _, issue := range result.Issues {
-			fmt.Printf("  %s\n", issue.String())
-		}
-		fmt.Println()
-	}
-
-	// Write files
-	if err := result.WriteFiles(flags.output); err != nil {
-		return fmt.Errorf("writing files: %w", err)
-	}
-
-	// Print generated files
-	fmt.Printf("Generated Files (%d):\n", len(result.Files))
-	for _, file := range result.Files {
-		fmt.Printf("  - %s/%s (%d bytes)\n", flags.output, file.Name, len(file.Content))
-	}
-	fmt.Println()
-
-	// Print summary
-	if result.Success {
-		fmt.Printf("✓ Generation successful")
-		if result.InfoCount > 0 || result.WarningCount > 0 {
-			fmt.Printf(" (%d info, %d warnings)", result.InfoCount, result.WarningCount)
-		}
-		fmt.Println()
-	} else {
-		fmt.Printf("✗ Generation completed with %d critical issue(s)", result.CriticalCount)
-		if result.WarningCount > 0 {
-			fmt.Printf(", %d warning(s)", result.WarningCount)
-		}
-		fmt.Println()
-		os.Exit(1)
-	}
-
-	return nil
-}
-
-// fixFlags contains flags for the fix command
-type fixFlags struct {
-	output string
-	infer  bool
-	quiet  bool
-}
-
-// setupFixFlags creates and configures a FlagSet for the fix command.
-// Returns the FlagSet and a fixFlags struct with bound flag variables.
-func setupFixFlags() (*flag.FlagSet, *fixFlags) {
-	fs := flag.NewFlagSet("fix", flag.ContinueOnError)
-	flags := &fixFlags{}
-
-	fs.StringVar(&flags.output, "o", "", "output file path (default: stdout)")
-	fs.StringVar(&flags.output, "output", "", "output file path (default: stdout)")
-	fs.BoolVar(&flags.infer, "infer", false, "infer parameter types from naming conventions")
-	fs.BoolVar(&flags.quiet, "q", false, "quiet mode: only output the document, no diagnostic messages")
-	fs.BoolVar(&flags.quiet, "quiet", false, "quiet mode: only output the document, no diagnostic messages")
-
-	fs.Usage = func() {
-		cliutil.Writef(fs.Output(), "Usage: oastools fix [flags] <file|url|->\n\n")
-		cliutil.Writef(fs.Output(), "Apply automatic fixes to common OpenAPI specification issues.\n\n")
-		cliutil.Writef(fs.Output(), "Flags:\n")
-		fs.PrintDefaults()
-		cliutil.Writef(fs.Output(), "\nSupported Fixes:\n")
-		cliutil.Writef(fs.Output(), "  - Missing path parameters: Adds Parameter objects for path template\n")
-		cliutil.Writef(fs.Output(), "    variables that are not declared in the operation's parameters list.\n")
-		cliutil.Writef(fs.Output(), "    Default type is 'string'. Use --infer for smart type inference.\n")
-		cliutil.Writef(fs.Output(), "\nType Inference (--infer):\n")
-		cliutil.Writef(fs.Output(), "  - Names ending in 'id', 'Id', 'ID' -> integer\n")
-		cliutil.Writef(fs.Output(), "  - Names containing 'uuid', 'guid' -> string with format uuid\n")
-		cliutil.Writef(fs.Output(), "  - All other names -> string\n")
-		cliutil.Writef(fs.Output(), "\nExamples:\n")
-		cliutil.Writef(fs.Output(), "  oastools fix openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools fix -o fixed.yaml openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  oastools fix --infer openapi.yaml\n")
-		cliutil.Writef(fs.Output(), "  cat openapi.yaml | oastools fix -q - > fixed.yaml\n")
-		cliutil.Writef(fs.Output(), "\nPipelining:\n")
-		cliutil.Writef(fs.Output(), "  oastools fix -q api.yaml | oastools validate -q -\n")
-		cliutil.Writef(fs.Output(), "  oastools fix -q --infer api.yaml | oastools convert -q -t 3.1.0 -\n")
-		cliutil.Writef(fs.Output(), "\nNotes:\n")
-		cliutil.Writef(fs.Output(), "  - Use 'oastools validate' to see validation errors before fixing\n")
-		cliutil.Writef(fs.Output(), "  - The fix command always applies all available fixes\n")
-		cliutil.Writef(fs.Output(), "  - Output preserves the original format (JSON or YAML)\n")
-		cliutil.Writef(fs.Output(), "\nExit Codes:\n")
-		cliutil.Writef(fs.Output(), "  0    Fixes applied successfully (or no fixes needed)\n")
-		cliutil.Writef(fs.Output(), "  1    Failed to parse or fix the specification\n")
-	}
-
-	return fs, flags
-}
-
-func handleFix(args []string) error {
-	fs, flags := setupFixFlags()
-
-	if err := fs.Parse(args); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return nil
-		}
-		return err
-	}
-
-	if fs.NArg() != 1 {
-		fs.Usage()
-		return fmt.Errorf("fix command requires exactly one file path, URL, or '-' for stdin")
-	}
-
-	specPath := fs.Arg(0)
-
-	// Create fixer with options
-	f := fixer.New()
-	f.InferTypes = flags.infer
-
-	// Fix the file, URL, or stdin with timing
-	startTime := time.Now()
-	var result *fixer.FixResult
-	var err error
-
-	if specPath == StdinFilePath {
-		// Read from stdin
-		p := parser.New()
-		parseResult, err := p.ParseReader(os.Stdin)
-		if err != nil {
-			return fmt.Errorf("parsing stdin: %w", err)
-		}
-		result, err = f.FixParsed(*parseResult)
-		if err != nil {
-			return fmt.Errorf("fixing from stdin: %w", err)
-		}
-	} else {
-		result, err = f.Fix(specPath)
-		if err != nil {
-			return fmt.Errorf("fixing file: %w", err)
-		}
-	}
-	totalTime := time.Since(startTime)
-
-	// Print diagnostic messages (to stderr to keep stdout clean for pipelining)
-	if !flags.quiet {
-		cliutil.Writef(os.Stderr, "OpenAPI Specification Fixer\n")
-		cliutil.Writef(os.Stderr, "===========================\n\n")
-		cliutil.Writef(os.Stderr, "oastools version: %s\n", oastools.Version())
-		if specPath == StdinFilePath {
-			cliutil.Writef(os.Stderr, "Specification: <stdin>\n")
-		} else {
-			cliutil.Writef(os.Stderr, "Specification: %s\n", specPath)
-		}
-		cliutil.Writef(os.Stderr, "OAS Version: %s\n", result.SourceVersion)
-		cliutil.Writef(os.Stderr, "Paths: %d\n", result.Stats.PathCount)
-		cliutil.Writef(os.Stderr, "Operations: %d\n", result.Stats.OperationCount)
-		cliutil.Writef(os.Stderr, "Schemas: %d\n", result.Stats.SchemaCount)
-		cliutil.Writef(os.Stderr, "Total Time: %v\n\n", totalTime)
-
-		// Print fixes applied
-		if result.HasFixes() {
-			cliutil.Writef(os.Stderr, "Fixes Applied (%d):\n", result.FixCount)
-			for _, fix := range result.Fixes {
-				cliutil.Writef(os.Stderr, "  - [%s] %s: %s\n", fix.Type, fix.Path, fix.Description)
-			}
-			cliutil.Writef(os.Stderr, "\n")
-		}
-
-		// Print summary
-		if result.HasFixes() {
-			cliutil.Writef(os.Stderr, "✓ Applied %d fix(es)\n", result.FixCount)
-		} else {
-			cliutil.Writef(os.Stderr, "✓ No fixes needed - specification is already valid\n")
-		}
-	}
-
-	// Write output
-	data, err := marshalDocument(result.Document, result.SourceFormat)
-	if err != nil {
-		return fmt.Errorf("marshaling fixed document: %w", err)
-	}
-
-	if flags.output != "" {
-		if err := os.WriteFile(flags.output, data, 0600); err != nil {
-			return fmt.Errorf("writing output file: %w", err)
-		}
-		if !flags.quiet {
-			cliutil.Writef(os.Stderr, "\nOutput written to: %s\n", flags.output)
-		}
-	} else {
-		// Write to stdout
-		if _, err = os.Stdout.Write(data); err != nil {
-			return fmt.Errorf("writing fixed document to stdout: %w", err)
-		}
-	}
-
-	return nil
 }
 
 func printUsage() {
